+++ src/renderer/panel/main.tsx
@@ -1,7 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { Panel } from './Panel'
+import { setLocale, detectLocale } from '../../shared/i18n'
 import './panel.css'
+
+setLocale(detectLocale())
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>